buil(app): create docker and docker-compose files, use postgres db instead of sqlite
