+++ .env.example
@@ -1,0 +1,5 @@
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_DB=etna
+POSTGRES_PORT=5432
+POSTGRES_HOST=localhost
