+++ cmd/main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 
-	_ "github.com/mattn/go-sqlite3"
 	"github.com/robfig/cron"
 	"github.com/spf13/viper"
 
@@ -28,6 +27,7 @@ func loadConfig() {
 	viper.SetConfigName("config")
 	viper.SetConfigType("yaml")
 	viper.AddConfigPath("./config")
+	viper.AutomaticEnv()
 	err := viper.ReadInConfig()
 	if err != nil {
 		panic(fmt.Errorf("fatal error config file: %w", err))
