+++ docker-compose.yml
@@ -1,0 +1,45 @@
+version: '3.4'
+
+services:
+  postgres:
+    image:  postgres:13-alpine
+    networks: [ app ]
+    environment:
+      - POSTGRES_USER=$POSTGRES_USER
+      - POSTGRES_PASSWORD=$POSTGRES_PASSWORD
+      - POSTGRES_DB=$POSTGRES_DB
+#    volumes:
+#      - ./data/postgresql:/var/lib/postgresql/data
+
+  migrate:
+    image: migrate/migrate
+    networks: [ app ]
+    environment:
+      - POSTGRES_USER=$POSTGRES_USER
+      - POSTGRES_PASSWORD=$POSTGRES_PASSWORD
+      - POSTGRES_DB=$POSTGRES_DB
+    volumes:
+      - ./database/migrations:/migrations
+    command: ["-path", "/migrations", "-database",  "postgres://$POSTGRES_USER:$POSTGRES_PASSWORD@postgres:5432/$POSTGRES_DB?sslmode=disable", "up"]
+    links:
+      - postgres
+    depends_on:
+      - postgres
+
+  app:
+    build: .
+    restart: always
+    networks: [ app ]
+    depends_on:
+      - postgres
+    links:
+      - postgres
+    environment:
+      - POSTGRES_USER=$POSTGRES_USER
+      - POSTGRES_PASSWORD=$POSTGRES_PASSWORD
+      - POSTGRES_DB=$POSTGRES_DB
+      - POSTGRES_HOST=$POSTGRES_HOST
+      - POSTGRES_PORT=$POSTGRES_PORT
+
+networks:
+  app:
