+++ go.mod
@@ -4,7 +4,7 @@ go 1.18
 
 require (
 	github.com/bwmarrin/discordgo v0.26.0
-	github.com/mattn/go-sqlite3 v1.14.15
+	github.com/lib/pq v1.10.6
 	github.com/maxatome/go-testdeep v1.12.0
 	github.com/robfig/cron v1.2.0
 	github.com/spf13/viper v1.12.0
